Split NyxidChat endpoints into focused partial files
Extract NyxIdChatEndpoints.cs (1566 lines) into responsibility-scoped
partials: Streaming, Relay, ConnectedServices. Pull the
subscribe/dispatch/timeout pattern shared by stream + approve into
NyxIdChatStreamingRunner. Move relay-only logic into NyxIdRelay{Models,
Replies, WorkflowCards}. Behavior is preserved; the new GetContentType
helper carries forward the content_type/type fallback added on the
remote branch.
