+++ agents/Aevatar.GAgents.NyxidChat/NyxIdChatEndpoints.cs
@@ -1,31 +1,20 @@
-using System.Security.Cryptography;
 using System.IdentityModel.Tokens.Jwt;
-using System.Text;
-using System.Text.Json;
-using Aevatar.AI.Abstractions;
 using Aevatar.AI.Abstractions.LLMProviders;
 using Aevatar.AI.ToolProviders.NyxId;
-using Aevatar.CQRS.Core.Abstractions.Commands;
 using Aevatar.Foundation.Abstractions;
 using Aevatar.Foundation.Abstractions.Streaming;
-using Aevatar.Workflow.Application.Abstractions.Runs;
-using Google.Protobuf.WellKnownTypes;
+using Aevatar.Studio.Application.Studio.Abstractions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.Extensions.Caching.Memory;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Aevatar.Studio.Application.Studio.Abstractions;
 using Microsoft.Extensions.Logging;
 
 namespace Aevatar.GAgents.NyxidChat;
 
-public static class NyxIdChatEndpoints
+public static partial class NyxIdChatEndpoints
 {
-    private const string RelayCardActionContentType = "card_action";
-
     public static IEndpointRouteBuilder MapNyxIdChatEndpoints(this IEndpointRouteBuilder app)
     {
         var group = app.MapGroup("/api/scopes").WithTags("NyxIdChat");
@@ -125,216 +114,6 @@ public static class NyxIdChatEndpoints
         }
     }
 
-    private static async Task HandleStreamMessageAsync(
-        HttpContext http,
-        string scopeId,
-        string actorId,
-        NyxIdChatStreamRequest request,
-        [FromServices] IActorRuntime actorRuntime,
-        [FromServices] IActorEventSubscriptionProvider subscriptionProvider,
-        [FromServices] ILoggerFactory loggerFactory,
-        CancellationToken ct)
-    {
-        var logger = loggerFactory.CreateLogger("Aevatar.NyxId.Chat.Endpoints");
-        var writer = new NyxIdChatSseWriter(http.Response);
-
-        try
-        {
-            var accessToken = ExtractBearerToken(http);
-            if (string.IsNullOrWhiteSpace(accessToken))
-            {
-                http.Response.StatusCode = StatusCodes.Status401Unauthorized;
-                return;
-            }
-
-            var prompt = request.Prompt?.Trim() ?? string.Empty;
-            if (string.IsNullOrWhiteSpace(prompt) && request.InputParts is not { Count: > 0 })
-            {
-                http.Response.StatusCode = StatusCodes.Status400BadRequest;
-                return;
-            }
-
-            // Get or create the actor
-            var actor = await actorRuntime.GetAsync(actorId)
-                        ?? await actorRuntime.CreateAsync<NyxIdChatGAgent>(actorId, ct);
-
-            // Set up SSE response
-            await writer.StartAsync(ct);
-
-            var messageId = Guid.NewGuid().ToString("N");
-            await writer.WriteRunStartedAsync(actorId, ct);
-
-            // Subscribe to actor events and map to SSE frames
-            var tcs = new TaskCompletionSource<string>(TaskCreationOptions.RunContinuationsAsynchronously);
-            using var ctr = ct.Register(() => tcs.TrySetCanceled());
-
-            await using var subscription = await subscriptionProvider.SubscribeAsync<EventEnvelope>(
-                actor.Id,
-                async envelope =>
-                {
-                    try
-                    {
-                        var terminalFrame = await MapAndWriteEventAsync(envelope, messageId, writer);
-                        if (!string.IsNullOrWhiteSpace(terminalFrame))
-                            tcs.TrySetResult(terminalFrame);
-                    }
-                    catch (Exception ex)
-                    {
-                        tcs.TrySetException(ex);
-                    }
-                },
-                ct);
-
-            // Build and dispatch ChatRequestEvent to the actor
-            var chatRequest = new ChatRequestEvent
-            {
-                Prompt = prompt,
-                SessionId = request.SessionId ?? messageId,
-                ScopeId = scopeId,
-            };
-            if (request.InputParts is { Count: > 0 })
-            {
-                foreach (var part in request.InputParts)
-                    chatRequest.InputParts.Add(part.ToProto());
-            }
-            chatRequest.Metadata[LLMRequestMetadataKeys.NyxIdAccessToken] = accessToken;
-            if (TryExtractRefreshToken(http) is { } refreshToken)
-                chatRequest.Metadata[LLMRequestMetadataKeys.NyxIdRefreshToken] = refreshToken;
-            chatRequest.Metadata["scope_id"] = scopeId;
-            await InjectUserConfigMetadataAsync(http, chatRequest.Metadata, ct);
-            await InjectUserMemoryAsync(http, chatRequest.Metadata, ct);
-            await InjectConnectedServicesAsync(http, accessToken, chatRequest.Metadata, ct);
-
-            var envelope = new EventEnvelope
-            {
-                Id = Guid.NewGuid().ToString("N"),
-                Timestamp = Timestamp.FromDateTimeOffset(DateTimeOffset.UtcNow),
-                Payload = Any.Pack(chatRequest),
-                Route = new EnvelopeRoute
-                {
-                    Direct = new DirectRoute { TargetActorId = actor.Id },
-                },
-            };
-
-            await actor.HandleEventAsync(envelope, ct);
-
-            // Wait for completion or timeout
-            var completedTask = await Task.WhenAny(tcs.Task, Task.Delay(120_000, ct));
-
-            if (completedTask == tcs.Task)
-            {
-                if (tcs.Task.IsFaulted)
-                {
-                    await writer.WriteRunErrorAsync(
-                        "The chat request failed before completion. Please try again.",
-                        CancellationToken.None);
-                }
-                else if (string.Equals(tcs.Task.Result, "TEXT_MESSAGE_END", StringComparison.Ordinal))
-                {
-                    await writer.WriteRunFinishedAsync(CancellationToken.None);
-                }
-            }
-            else
-            {
-                await writer.WriteRunErrorAsync("Request timed out.", CancellationToken.None);
-            }
-        }
-        catch (OperationCanceledException)
-        {
-            // Client disconnected
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, "NyxID chat stream failed for actor {ActorId}", actorId);
-            if (!writer.Started)
-            {
-                http.Response.StatusCode = StatusCodes.Status500InternalServerError;
-                return;
-            }
-
-            await writer.WriteRunErrorAsync(
-                "The chat request failed. Please try again.",
-                CancellationToken.None);
-        }
-    }
-
-    /// <summary>
-    /// Maps AI event envelope payloads to NyxIdChat SSE frames.
-    /// </summary>
-    private static async ValueTask<string?> MapAndWriteEventAsync(
-        EventEnvelope envelope,
-        string messageId,
-        NyxIdChatSseWriter writer)
-    {
-        var payload = envelope.Payload;
-        if (payload is null)
-            return null;
-
-        if (payload.Is(TextMessageStartEvent.Descriptor))
-        {
-            await writer.WriteTextStartAsync(messageId, CancellationToken.None);
-        }
-        else if (payload.Is(TextMessageContentEvent.Descriptor))
-        {
-            var evt = payload.Unpack<TextMessageContentEvent>();
-            if (!string.IsNullOrEmpty(evt.Delta))
-                await writer.WriteTextDeltaAsync(evt.Delta, CancellationToken.None);
-        }
-        else if (payload.Is(ToolCallEvent.Descriptor))
-        {
-            var evt = payload.Unpack<ToolCallEvent>();
-            await writer.WriteToolCallStartAsync(evt.ToolName, evt.CallId, CancellationToken.None);
-        }
-        else if (payload.Is(ToolResultEvent.Descriptor))
-        {
-            var evt = payload.Unpack<ToolResultEvent>();
-            await writer.WriteToolCallEndAsync(evt.CallId, evt.ResultJson, CancellationToken.None);
-        }
-        else if (payload.Is(ToolApprovalRequestEvent.Descriptor))
-        {
-            var evt = payload.Unpack<ToolApprovalRequestEvent>();
-            await writer.WriteToolApprovalRequestAsync(
-                evt.RequestId, evt.ToolName, evt.ToolCallId,
-                evt.ArgumentsJson, evt.IsDestructive, evt.TimeoutSeconds,
-                CancellationToken.None);
-        }
-        else if (payload.Is(MediaContentEvent.Descriptor))
-        {
-            var evt = payload.Unpack<MediaContentEvent>();
-            await writer.WriteMediaContentAsync(evt, CancellationToken.None);
-        }
-        else if (payload.Is(TextMessageEndEvent.Descriptor))
-        {
-            var evt = payload.Unpack<TextMessageEndEvent>();
-
-            // Check for LLM error markers
-            if (!string.IsNullOrEmpty(evt.Content))
-            {
-                const string llmErrorPrefix = "[[AEVATAR_LLM_ERROR]]";
-                const string llmFailedPrefix = "LLM request failed:";
-                if (evt.Content.StartsWith(llmErrorPrefix, StringComparison.Ordinal))
-                {
-                    await writer.WriteRunErrorAsync(
-                        ClassifyError(evt.Content[llmErrorPrefix.Length..].Trim()), CancellationToken.None);
-                    return "RUN_ERROR";
-                }
-
-                if (evt.Content.StartsWith(llmFailedPrefix, StringComparison.Ordinal))
-                {
-                    await writer.WriteRunErrorAsync(
-                        ClassifyError(evt.Content.Trim()),
-                        CancellationToken.None);
-                    return "RUN_ERROR";
-                }
-            }
-
-            await writer.WriteTextEndAsync(messageId, CancellationToken.None);
-            return "TEXT_MESSAGE_END";
-        }
-
-        return null;
-    }
-
     private static async Task<IResult> HandleDeleteConversationAsync(
         HttpContext http,
         string scopeId,
@@ -347,137 +126,6 @@ public static class NyxIdChatEndpoints
         await actorStore.RemoveActorAsync(scopeId, NyxIdChatServiceDefaults.GAgentTypeName, actorId, ct);
         return Results.Ok();
     }
-
-    /// <summary>
-    /// Handles tool approval decisions from the frontend.
-    /// Opens an SSE connection to stream the continuation chat response.
-    /// </summary>
-    private static async Task HandleApproveAsync(
-        HttpContext http,
-        string scopeId,
-        string actorId,
-        NyxIdApprovalRequest request,
-        [FromServices] IActorRuntime actorRuntime,
-        [FromServices] IActorEventSubscriptionProvider subscriptionProvider,
-        [FromServices] ILoggerFactory loggerFactory,
-        CancellationToken ct)
-    {
-        var logger = loggerFactory.CreateLogger("Aevatar.NyxId.Chat.Endpoints");
-        var writer = new NyxIdChatSseWriter(http.Response);
-
-        try
-        {
-            var accessToken = ExtractBearerToken(http);
-            if (string.IsNullOrWhiteSpace(accessToken))
-            {
-                http.Response.StatusCode = StatusCodes.Status401Unauthorized;
-                return;
-            }
-
-            if (string.IsNullOrWhiteSpace(request.RequestId))
-            {
-                http.Response.StatusCode = StatusCodes.Status400BadRequest;
-                return;
-            }
-
-            var actor = await actorRuntime.GetAsync(actorId);
-            if (actor == null)
-            {
-                http.Response.StatusCode = StatusCodes.Status404NotFound;
-                return;
-            }
-
-            await writer.StartAsync(ct);
-            await writer.WriteRunStartedAsync(actorId, ct);
-
-            var messageId = Guid.NewGuid().ToString("N");
-            var tcs = new TaskCompletionSource<string>(TaskCreationOptions.RunContinuationsAsynchronously);
-            using var ctr = ct.Register(() => tcs.TrySetCanceled());
-
-            await using var subscription = await subscriptionProvider.SubscribeAsync<EventEnvelope>(
-                actor.Id,
-                async envelope =>
-                {
-                    try
-                    {
-                        var terminalFrame = await MapAndWriteEventAsync(envelope, messageId, writer);
-                        if (!string.IsNullOrWhiteSpace(terminalFrame))
-                            tcs.TrySetResult(terminalFrame);
-                    }
-                    catch (Exception ex)
-                    {
-                        tcs.TrySetException(ex);
-                    }
-                },
-                ct);
-
-            // Send the approval decision to the actor
-            var decisionEvent = new ToolApprovalDecisionEvent
-            {
-                RequestId = request.RequestId,
-                SessionId = request.SessionId ?? scopeId,
-                Approved = request.Approved,
-                Reason = request.Reason ?? string.Empty,
-            };
-
-            var envelope = new EventEnvelope
-            {
-                Id = Guid.NewGuid().ToString("N"),
-                Timestamp = Timestamp.FromDateTimeOffset(DateTimeOffset.UtcNow),
-                Payload = Any.Pack(decisionEvent),
-                Route = new EnvelopeRoute
-                {
-                    Direct = new DirectRoute { TargetActorId = actor.Id },
-                },
-            };
-
-            await actor.HandleEventAsync(envelope, ct);
-
-            // Wait for the continuation chat to complete (or timeout)
-            var completedTask = await Task.WhenAny(tcs.Task, Task.Delay(120_000, ct));
-
-            if (completedTask == tcs.Task)
-            {
-                if (tcs.Task.IsFaulted)
-                {
-                    await writer.WriteRunErrorAsync(
-                        "The approval continuation failed before completion. Please try again.",
-                        CancellationToken.None);
-                }
-                else if (string.Equals(tcs.Task.Result, "TEXT_MESSAGE_END", StringComparison.Ordinal))
-                {
-                    await writer.WriteRunFinishedAsync(CancellationToken.None);
-                }
-            }
-            else
-            {
-                await writer.WriteRunErrorAsync("Approval continuation timed out.", CancellationToken.None);
-            }
-        }
-        catch (OperationCanceledException)
-        {
-            // Client disconnected
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, "NyxID approval stream failed for actor {ActorId}", actorId);
-            if (!writer.Started)
-            {
-                http.Response.StatusCode = StatusCodes.Status500InternalServerError;
-                return;
-            }
-
-            await writer.WriteRunErrorAsync(
-                "The approval continuation failed. Please try again.",
-                CancellationToken.None);
-        }
-    }
-
-    public sealed record NyxIdApprovalRequest(
-        string? RequestId,
-        bool Approved = true,
-        string? Reason = null,
-        string? SessionId = null);
 
     private static async Task InjectUserConfigMetadataAsync(
         HttpContext http,
@@ -540,135 +188,6 @@ public static class NyxIdChatEndpoints
         }
     }
 
-    private static async Task InjectConnectedServicesAsync(
-        HttpContext http,
-        string accessToken,
-        IDictionary<string, string> metadata,
-        CancellationToken ct)
-    {
-        var client = http.RequestServices.GetService<NyxIdApiClient>();
-        if (client is null)
-            return;
-
-        var logger = http.RequestServices.GetService<ILoggerFactory>()
-            ?.CreateLogger("Aevatar.NyxId.Chat.ConnectedServices");
-
-        try
-        {
-            var memCache = http.RequestServices.GetService<IMemoryCache>();
-            var cacheKey = $"nyxid:services:{ComputeTokenHash(accessToken)}";
-
-            string? servicesJson = null;
-            if (memCache is not null)
-                servicesJson = memCache.Get<string>(cacheKey);
-
-            if (servicesJson is null)
-            {
-                servicesJson = await client.DiscoverProxyServicesAsync(accessToken, ct);
-                memCache?.Set(cacheKey, servicesJson, TimeSpan.FromSeconds(60));
-            }
-
-            var specSource = http.RequestServices.GetService<IConnectedServiceSpecSource>();
-            var context = await BuildConnectedServicesContextAsync(servicesJson, specSource, accessToken, ct);
-            if (!string.IsNullOrWhiteSpace(context))
-                metadata[LLMRequestMetadataKeys.ConnectedServicesContext] = context;
-        }
-        catch (Exception ex)
-        {
-            logger?.LogWarning(ex, "Failed to load connected services context — continuing without capability context");
-        }
-    }
-
-    internal static async Task<string> BuildConnectedServicesContextAsync(
-        string servicesJson,
-        IConnectedServiceSpecSource? specSource,
-        string accessToken,
-        CancellationToken ct)
-    {
-        var sb = new StringBuilder();
-        sb.AppendLine("<connected-services>");
-        sb.AppendLine("Your capabilities based on connected services:");
-
-        var hintRequests = new List<ServiceHintRequest>();
-
-        try
-        {
-            using var doc = JsonDocument.Parse(servicesJson);
-            var root = doc.RootElement;
-
-            JsonElement items = root;
-            if (root.ValueKind == JsonValueKind.Object)
-            {
-                if (root.TryGetProperty("services", out var svc))
-                    items = svc;
-                else if (root.TryGetProperty("data", out var data))
-                    items = data;
-            }
-
-            if (items.ValueKind == JsonValueKind.Array)
-            {
-                foreach (var item in items.EnumerateArray())
-                {
-                    var serviceId = item.TryGetProperty("id", out var id) ? id.GetString()
-                                  : item.TryGetProperty("service_id", out var sid) ? sid.GetString()
-                                  : null;
-                    var slug = item.TryGetProperty("slug", out var s) ? s.GetString() : null;
-                    var name = item.TryGetProperty("name", out var n) ? n.GetString()
-                             : item.TryGetProperty("label", out var l) ? l.GetString()
-                             : slug;
-                    var baseUrl = item.TryGetProperty("endpoint_url", out var e) ? e.GetString()
-                                : item.TryGetProperty("base_url", out var b) ? b.GetString()
-                                : null;
-                    var openapiUrl = item.TryGetProperty("openapi_url", out var oa) ? oa.GetString() : null;
-
-                    if (string.IsNullOrWhiteSpace(slug)) continue;
-                    hintRequests.Add(new ServiceHintRequest(slug, serviceId, name, openapiUrl));
-
-                    sb.Append($"- **{name ?? slug}** (slug: `{slug}`)");
-                    if (!string.IsNullOrWhiteSpace(baseUrl))
-                        sb.Append($" — base: {baseUrl}");
-                    sb.AppendLine();
-                }
-            }
-        }
-        catch
-        {
-            // Parse failure — return what we have
-        }
-
-        if (hintRequests.Count == 0)
-        {
-            sb.AppendLine("No services connected yet. Use nyxid_catalog to browse and connect services.");
-        }
-
-        sb.AppendLine("Use nyxid_proxy with slug + path to call any service. Use code_execute for sandbox.");
-        sb.AppendLine("</connected-services>");
-
-        string hints;
-        if (specSource is not null && !string.IsNullOrWhiteSpace(accessToken))
-        {
-            hints = await NyxIdServiceApiHints.BuildHintsSectionAsync(hintRequests, specSource, accessToken, ct);
-        }
-        else
-        {
-            hints = NyxIdServiceApiHints.BuildHintsSection(hintRequests.Select(r => r.Slug));
-        }
-
-        if (!string.IsNullOrEmpty(hints))
-        {
-            sb.AppendLine();
-            sb.Append(hints);
-        }
-
-        return sb.ToString();
-    }
-
-    private static string ComputeTokenHash(string token)
-    {
-        var bytes = SHA256.HashData(Encoding.UTF8.GetBytes(token));
-        return Convert.ToHexStringLower(bytes)[..16]; // Short hash for cache key
-    }
-
     private static string? ExtractBearerToken(HttpContext http)
     {
         var authHeader = http.Request.Headers.Authorization.FirstOrDefault();
@@ -724,853 +243,4 @@ public static class NyxIdChatEndpoints
             return null;
         }
     }
-
-    public sealed record NyxIdChatStreamRequest(
-        string? Prompt,
-        string? SessionId = null,
-        IReadOnlyList<ContentPartDto>? InputParts = null);
-
-    public sealed record ContentPartDto(
-        string Type,
-        string? Text = null,
-        string? DataBase64 = null,
-        string? MediaType = null,
-        string? Uri = null,
-        string? Name = null)
-    {
-        public ChatContentPart ToProto() => new()
-        {
-            Kind = Type?.ToLowerInvariant() switch
-            {
-                "image" => ChatContentPartKind.Image,
-                "audio" => ChatContentPartKind.Audio,
-                "video" => ChatContentPartKind.Video,
-                "text" => ChatContentPartKind.Text,
-                _ => ChatContentPartKind.Unspecified,
-            },
-            Text = Text ?? string.Empty,
-            DataBase64 = DataBase64 ?? string.Empty,
-            MediaType = MediaType ?? string.Empty,
-            Uri = Uri ?? string.Empty,
-            Name = Name ?? string.Empty,
-        };
-    }
-
-    // ─── NyxID Channel Bot Relay ───
-
-    /// <summary>
-    /// Receives forwarded platform messages from NyxID Channel Bot Relay.
-    /// Validates the Nyx relay JWT, durably dispatches the inbound turn, returns 202,
-    /// and sends the platform reply asynchronously through Nyx channel-relay/reply.
-    /// </summary>
-    private static async Task<IResult> HandleRelayWebhookAsync(
-        HttpContext http,
-        [FromServices] IActorRuntime actorRuntime,
-        [FromServices] IActorEventSubscriptionProvider subscriptionProvider,
-        [FromServices] IGAgentActorStore actorStore,
-        [FromServices] NyxIdRelayOptions relayOptions,
-        [FromServices] NyxRelayJwtValidator relayJwtValidator,
-        [FromServices] NyxIdApiClient nyxClient,
-        [FromServices] ILoggerFactory loggerFactory,
-        CancellationToken ct)
-    {
-        var logger = loggerFactory.CreateLogger("Aevatar.NyxId.Chat.Relay");
-
-        try
-        {
-            // ─── Parse payload ───
-            RelayMessage? message;
-            try
-            {
-                message = await http.Request.ReadFromJsonAsync<RelayMessage>(RelayJsonOptions, ct);
-            }
-            catch (JsonException ex)
-            {
-                logger.LogWarning(ex, "Failed to parse relay payload");
-                return Results.BadRequest(new { error = "invalid_relay_payload" });
-            }
-
-            if (message is null)
-                return Results.BadRequest(new { error = "missing_relay_payload" });
-            if (string.IsNullOrWhiteSpace(message.MessageId))
-                return Results.BadRequest(new { error = "missing_message_id" });
-            var contentType = GetRelayContentType(message.Content);
-            var contentText = NormalizeOptional(message.Content?.Text);
-            if (contentText is null)
-                return Results.Accepted(value: new { status = "ignored", reason = "empty_text" });
-
-            // ─── Auth ───
-            var userToken = http.Request.Headers["X-NyxID-User-Token"].FirstOrDefault();
-            if (string.IsNullOrWhiteSpace(userToken))
-            {
-                logger.LogWarning("Relay callback missing X-NyxID-User-Token header");
-                return Results.Unauthorized();
-            }
-
-            var validation = await relayJwtValidator.ValidateAsync(userToken, ct);
-            if (!validation.Succeeded || validation.Principal is null || string.IsNullOrWhiteSpace(validation.Subject))
-            {
-                logger.LogWarning("Relay callback JWT validation failed: {Error}", validation.Error);
-                return Results.Unauthorized();
-            }
-
-            if (!string.IsNullOrWhiteSpace(message.Agent?.ApiKeyId) &&
-                !string.Equals(message.Agent.ApiKeyId, validation.RelayApiKeyId, StringComparison.Ordinal))
-            {
-                logger.LogWarning(
-                    "Relay callback agent mismatch: payload_agent_api_key_id={PayloadApiKeyId}, token_agent_api_key_id={TokenApiKeyId}",
-                    message.Agent.ApiKeyId,
-                    validation.RelayApiKeyId);
-                return Results.Unauthorized();
-            }
-
-            http.User = validation.Principal;
-            var scopeId = validation.Subject!;
-
-            if (string.Equals(contentType, RelayCardActionContentType, StringComparison.Ordinal))
-            {
-                if (await TryHandleRelayWorkflowCardActionAsync(http, message, logger, ct) is { } workflowResult)
-                    return workflowResult;
-
-                logger.LogInformation(
-                    "Ignored unsupported relay card action: message={MessageId}, conversation={ConversationId}",
-                    message.MessageId,
-                    message.Conversation?.Id ?? message.Conversation?.PlatformId);
-                return Results.Accepted(value: new
-                {
-                    status = "ignored",
-                    reason = "unsupported_card_action",
-                    message_id = message.MessageId,
-                });
-            }
-
-            // Note: config.json in chrono-storage cannot be read in relay flow because
-            // ChronoStorageCatalogBlobClient reads the Bearer token from Authorization header,
-            // which is not present on relay callbacks (token is in X-NyxID-User-Token instead).
-            // InjectUserConfigMetadataAsync will silently fall back to server defaults.
-
-            // ─── Resolve conversation ───
-            var platform = message.Platform ?? "unknown";
-            var conversationPlatformId = message.Conversation?.PlatformId ?? "unknown";
-            var conversationId = message.Conversation?.Id;
-            if (string.IsNullOrWhiteSpace(conversationId))
-                conversationId = $"{platform}-{conversationPlatformId}";
-
-            var actorId = $"nyxid-relay-{conversationId}";
-
-            logger.LogInformation(
-                "Relay message: platform={Platform}, conversation={ConversationId}, sender={Sender}",
-                platform, conversationId, message.Sender?.DisplayName);
-
-            // ─── Get or create actor ───
-            // Relay follows the same strict lifecycle contract as create:
-            // registry persistence via IGAgentActorStore is mandatory, and
-            // registry failures must surface instead of silently degrading.
-            var actor = await actorRuntime.GetAsync(actorId)
-                        ?? await actorRuntime.CreateAsync<NyxIdChatGAgent>(actorId, ct);
-            await actorStore.AddActorAsync(scopeId, NyxIdChatServiceDefaults.GAgentTypeName, actorId, ct);
-
-            // ─── Subscribe before dispatch so we can relay the response asynchronously ───
-            var responseTcs = new TaskCompletionSource<string>(TaskCreationOptions.RunContinuationsAsynchronously);
-            var relayReply = new RelayReplyAccumulator(relayOptions.MaxBufferedResponseChars);
-            var sessionId = !string.IsNullOrWhiteSpace(message.MessageId)
-                ? $"{conversationId}-{message.MessageId}"
-                : $"{conversationId}-{Guid.NewGuid():N}";
-
-            var subscription = await subscriptionProvider.SubscribeAsync<EventEnvelope>(
-                actor.Id,
-                envelope =>
-                {
-                    var payload = envelope.Payload;
-                    if (payload is null) return Task.CompletedTask;
-
-                    if (payload.Is(TextMessageContentEvent.Descriptor))
-                    {
-                        var evt = payload.Unpack<TextMessageContentEvent>();
-                        if (!string.Equals(evt.SessionId, sessionId, StringComparison.Ordinal))
-                            return Task.CompletedTask;
-                        relayReply.Append(evt.Delta);
-                    }
-                    else if (payload.Is(TextMessageEndEvent.Descriptor))
-                    {
-                        var evt = payload.Unpack<TextMessageEndEvent>();
-                        if (!string.Equals(evt.SessionId, sessionId, StringComparison.Ordinal))
-                            return Task.CompletedTask;
-
-                        if (TryExtractLlmError(evt.Content, out var extractedError))
-                        {
-                            relayReply.SetError(extractedError);
-                        }
-                        else if (relayReply.IsEmpty && !string.IsNullOrWhiteSpace(evt.Content))
-                        {
-                            relayReply.Append(evt.Content);
-                        }
-
-                        responseTcs.TrySetResult(relayReply.Snapshot());
-                    }
-
-                    return Task.CompletedTask;
-                },
-                CancellationToken.None);
-
-            // ─── Dispatch to actor ───
-            // SessionId = per-message unique ID (for idempotent retry),
-            // NOT conversationId (which is per-chat and would collide across messages).
-            var chatRequest = new ChatRequestEvent
-            {
-                Prompt = contentText,
-                SessionId = sessionId,
-                ScopeId = scopeId,
-            };
-            chatRequest.Metadata[LLMRequestMetadataKeys.NyxIdAccessToken] = userToken;
-            if (TryExtractRefreshToken(http) is { } refreshToken)
-                chatRequest.Metadata[LLMRequestMetadataKeys.NyxIdRefreshToken] = refreshToken;
-            chatRequest.Metadata["scope_id"] = scopeId;
-            chatRequest.Metadata["relay.platform"] = message.Platform ?? "";
-            chatRequest.Metadata["relay.sender"] = message.Sender?.DisplayName ?? "";
-            chatRequest.Metadata["relay.message_id"] = message.MessageId ?? "";
-            await InjectUserConfigMetadataAsync(http, chatRequest.Metadata, ct);
-            await InjectUserMemoryAsync(http, chatRequest.Metadata, ct);
-
-            var envelope = new EventEnvelope
-            {
-                Id = Guid.NewGuid().ToString("N"),
-                Timestamp = Timestamp.FromDateTimeOffset(DateTimeOffset.UtcNow),
-                Payload = Any.Pack(chatRequest),
-                Route = new EnvelopeRoute
-                {
-                    Direct = new DirectRoute { TargetActorId = actor.Id },
-                },
-            };
-
-            await actor.HandleEventAsync(envelope, ct);
-
-            var configuration = http.RequestServices.GetService<IConfiguration>();
-            _ = FinalizeRelayReplyAsync(
-                    subscription,
-                    responseTcs,
-                    relayReply,
-                    sessionId,
-                    message.MessageId!,
-                    userToken,
-                    chatRequest.Metadata,
-                    relayOptions,
-                    nyxClient,
-                    configuration,
-                    logger)
-                .ContinueWith(
-                    task => logger.LogError(task.Exception, "Relay background reply pipeline failed for session {SessionId}", sessionId),
-                    TaskContinuationOptions.OnlyOnFaulted);
-
-            return Results.Accepted(value: new
-            {
-                status = "accepted",
-                session_id = sessionId,
-                message_id = message.MessageId,
-            });
-        }
-        catch (OperationCanceledException)
-        {
-            return Results.StatusCode(499);
-        }
-        catch (InvalidOperationException)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, "Relay handler unexpected error");
-            return Results.StatusCode(StatusCodes.Status500InternalServerError);
-        }
-    }
-
-    /// <summary>Classify a technical LLM error into a user-friendly message.</summary>
-    private static string ClassifyError(string error)
-    {
-        if (error.Contains("403", StringComparison.Ordinal) ||
-            error.Contains("Forbidden", StringComparison.OrdinalIgnoreCase))
-            return "Sorry, I can't reach the AI service right now (403 Forbidden).";
-
-        if (error.Contains("401", StringComparison.Ordinal) ||
-            error.Contains("Unauthorized", StringComparison.OrdinalIgnoreCase) ||
-            error.Contains("authentication", StringComparison.OrdinalIgnoreCase))
-            return "Sorry, authentication with the AI service failed (401).";
-
-        if (error.Contains("429", StringComparison.Ordinal) ||
-            error.Contains("rate limit", StringComparison.OrdinalIgnoreCase) ||
-            error.Contains("too many", StringComparison.OrdinalIgnoreCase))
-            return "Sorry, the AI service is busy right now (429). Please wait a moment and try again.";
-
-        if (error.Contains("timeout", StringComparison.OrdinalIgnoreCase))
-            return "Sorry, the AI service took too long to respond. Please try again.";
-
-        if (error.Contains("model", StringComparison.OrdinalIgnoreCase) &&
-            error.Contains("not found", StringComparison.OrdinalIgnoreCase))
-            return "Sorry, the configured AI model is not available.";
-
-        return "Sorry, something went wrong while generating a response.";
-    }
-
-    private static bool TryExtractLlmError(string? content, out string error)
-    {
-        error = string.Empty;
-        if (string.IsNullOrWhiteSpace(content))
-            return false;
-
-        const string llmErrorPrefix = "[[AEVATAR_LLM_ERROR]]";
-        const string llmFailedPrefix = "LLM request failed:";
-        if (content.StartsWith(llmErrorPrefix, StringComparison.Ordinal))
-        {
-            error = content[llmErrorPrefix.Length..].Trim();
-            return true;
-        }
-
-        if (content.StartsWith(llmFailedPrefix, StringComparison.Ordinal))
-        {
-            error = content.Trim();
-            return true;
-        }
-
-        return false;
-    }
-
-    private static async Task FinalizeRelayReplyAsync(
-        IAsyncDisposable subscription,
-        TaskCompletionSource<string> responseTcs,
-        RelayReplyAccumulator relayReply,
-        string sessionId,
-        string messageId,
-        string relayToken,
-        Google.Protobuf.Collections.MapField<string, string> metadata,
-        NyxIdRelayOptions relayOptions,
-        NyxIdApiClient nyxClient,
-        IConfiguration? configuration,
-        ILogger logger)
-    {
-        await using var ownedSubscription = subscription;
-        using var timeoutCts = new CancellationTokenSource(TimeSpan.FromSeconds(relayOptions.ResponseTimeoutSeconds));
-
-        var completed = await Task.WhenAny(
-            responseTcs.Task,
-            Task.Delay(Timeout.InfiniteTimeSpan, timeoutCts.Token));
-
-        string replyText;
-        if (completed == responseTcs.Task && responseTcs.Task.IsCompletedSuccessfully)
-        {
-            replyText = responseTcs.Task.Result;
-            logger.LogInformation(
-                "Relay response completed: session={SessionId}, length={Length}",
-                sessionId,
-                replyText.Length);
-        }
-        else
-        {
-            var partial = relayReply.Snapshot();
-            logger.LogWarning(
-                "Relay response timed out: session={SessionId}, partial_length={Length}",
-                sessionId,
-                partial.Length);
-            replyText = partial.Length > 0
-                ? partial
-                : "Sorry, it's taking too long to respond. Please try again.";
-        }
-
-        if (relayReply.WasTruncated)
-        {
-            logger.LogWarning(
-                "Relay response buffer truncated: session={SessionId}, max_chars={MaxChars}",
-                sessionId,
-                relayReply.MaxChars);
-        }
-
-        var relayError = relayReply.GetErrorMessage();
-        if (!string.IsNullOrWhiteSpace(relayError))
-        {
-            logger.LogWarning(
-                "Relay LLM error surfaced through async reply: session={SessionId}, error={Error}",
-                sessionId,
-                relayError);
-            replyText = ClassifyError(relayError);
-            if (relayOptions.EnableDebugDiagnostics)
-            {
-                var diagnostic = BuildRelayDiagnostic(metadata, configuration, relayError);
-                replyText += $"\n\n[Debug]\n{diagnostic}";
-            }
-        }
-        else if (string.IsNullOrWhiteSpace(replyText))
-        {
-            replyText = "Sorry, I wasn't able to generate a response. Please try again.";
-        }
-
-        using var deliveryCts = new CancellationTokenSource(TimeSpan.FromSeconds(30));
-        var delivery = await nyxClient.SendChannelRelayTextReplyAsync(relayToken, messageId, replyText, deliveryCts.Token);
-        if (!delivery.Succeeded)
-        {
-            logger.LogError(
-                "Relay async reply delivery failed: session={SessionId}, messageId={MessageId}, detail={Detail}",
-                sessionId,
-                messageId,
-                delivery.Detail);
-            return;
-        }
-
-        logger.LogInformation(
-            "Relay async reply delivered: session={SessionId}, messageId={MessageId}, platformMessageId={PlatformMessageId}",
-            sessionId,
-            delivery.MessageId,
-            delivery.PlatformMessageId);
-    }
-
-    /// <summary>
-    /// Build diagnostic block for relay error replies. Only included when
-    /// <see cref="NyxIdRelayOptions.EnableDebugDiagnostics"/> is true.
-    /// </summary>
-    private static string BuildRelayDiagnostic(
-        Google.Protobuf.Collections.MapField<string, string> metadata,
-        IConfiguration? configuration,
-        string errorMessage)
-    {
-        var modelOverride = metadata.TryGetValue(LLMRequestMetadataKeys.ModelOverride, out var m) ? m : null;
-        var serverDefault = configuration?["Aevatar:NyxId:DefaultModel"] ?? "(OpenAIModel option)";
-        var route = metadata.TryGetValue(LLMRequestMetadataKeys.NyxIdRoutePreference, out var r)
-            && !string.IsNullOrWhiteSpace(r) ? r : "gateway";
-        var hasToken = metadata.ContainsKey(LLMRequestMetadataKeys.NyxIdAccessToken);
-        var scope = metadata.TryGetValue("scope_id", out var s) ? s : "<unknown>";
-
-        var model = !string.IsNullOrWhiteSpace(modelOverride)
-            ? $"{modelOverride} (from config.json)"
-            : $"server-default={serverDefault}";
-
-        var error = errorMessage.Length > 300 ? errorMessage[..300] + "..." : errorMessage;
-
-        return $"Model: {model}\nRoute: {route}\nScope: {scope}\nToken: {(hasToken ? "present" : "MISSING")}\nError: {error}";
-    }
-
-    private static async Task<IResult?> TryHandleRelayWorkflowCardActionAsync(
-        HttpContext http,
-        RelayMessage message,
-        ILogger logger,
-        CancellationToken ct)
-    {
-        if (!TryBuildRelayWorkflowResumeCommand(message, out var command))
-            return null;
-
-        var dispatchService = http.RequestServices.GetService<
-            ICommandDispatchService<WorkflowResumeCommand, WorkflowRunControlAcceptedReceipt, WorkflowRunControlStartError>>();
-        if (dispatchService is null)
-        {
-            logger.LogError(
-                "Workflow resume service unavailable for relay card action: message={MessageId}",
-                message.MessageId);
-            return Results.Json(
-                new
-                {
-                    error = "workflow_resume_service_unavailable",
-                    message = "Workflow resume service unavailable.",
-                },
-                statusCode: StatusCodes.Status503ServiceUnavailable);
-        }
-
-        var dispatch = await dispatchService.DispatchAsync(command!, ct);
-        if (!dispatch.Succeeded || dispatch.Receipt is null)
-        {
-            var error = dispatch.Error;
-            if (error is null)
-            {
-                return Results.Json(
-                    new
-                    {
-                        error = "workflow_resume_dispatch_failed",
-                        message = "Workflow control dispatch failed.",
-                    },
-                    statusCode: StatusCodes.Status503ServiceUnavailable);
-            }
-
-            return error.Code switch
-            {
-                WorkflowRunControlStartErrorCode.InvalidActorId => Results.Json(
-                    new { error = "invalid_actor_id", message = "actorId is required." },
-                    statusCode: StatusCodes.Status400BadRequest),
-                WorkflowRunControlStartErrorCode.InvalidRunId => Results.Json(
-                    new { error = "invalid_run_id", message = "runId is required." },
-                    statusCode: StatusCodes.Status400BadRequest),
-                WorkflowRunControlStartErrorCode.InvalidStepId => Results.Json(
-                    new { error = "invalid_step_id", message = "stepId is required." },
-                    statusCode: StatusCodes.Status400BadRequest),
-                WorkflowRunControlStartErrorCode.ActorNotFound => Results.Json(
-                    new { error = "actor_not_found", message = $"Actor '{error.ActorId}' not found." },
-                    statusCode: StatusCodes.Status404NotFound),
-                WorkflowRunControlStartErrorCode.ActorNotWorkflowRun => Results.Json(
-                    new
-                    {
-                        error = "actor_not_workflow_run",
-                        message = $"Actor '{error.ActorId}' is not a workflow run actor.",
-                    },
-                    statusCode: StatusCodes.Status409Conflict),
-                WorkflowRunControlStartErrorCode.RunBindingMissing => Results.Json(
-                    new
-                    {
-                        error = "run_binding_missing",
-                        message = $"Actor '{error.ActorId}' does not have a bound run id.",
-                    },
-                    statusCode: StatusCodes.Status409Conflict),
-                WorkflowRunControlStartErrorCode.RunBindingMismatch => Results.Json(
-                    new
-                    {
-                        error = "run_binding_mismatch",
-                        message =
-                            $"Actor '{error.ActorId}' is bound to run '{error.BoundRunId}', not '{error.RequestedRunId}'.",
-                    },
-                    statusCode: StatusCodes.Status409Conflict),
-                _ => Results.Json(
-                    new
-                    {
-                        error = "workflow_resume_dispatch_failed",
-                        message = "Workflow control dispatch failed.",
-                    },
-                    statusCode: StatusCodes.Status503ServiceUnavailable),
-            };
-        }
-
-        return Results.Accepted(value: new
-        {
-            status = "workflow_resume_accepted",
-            message_id = message.MessageId,
-            command_id = dispatch.Receipt.CommandId,
-        });
-    }
-
-    private static bool TryBuildRelayWorkflowResumeCommand(
-        RelayMessage message,
-        out WorkflowResumeCommand? command)
-    {
-        command = null;
-
-        if (!string.Equals(
-                GetRelayContentType(message.Content),
-                RelayCardActionContentType,
-                StringComparison.Ordinal))
-        {
-            return false;
-        }
-
-        var payload = NormalizeOptional(message.Content?.Text);
-        if (payload is null)
-            return false;
-
-        Dictionary<string, string> values;
-        try
-        {
-            using var document = JsonDocument.Parse(payload);
-            values = ExtractRelayCardActionValues(document.RootElement);
-        }
-        catch (JsonException)
-        {
-            return false;
-        }
-
-        if (!TryGetRequiredRelayValue(values, "actor_id", out var actorId) ||
-            !TryGetRequiredRelayValue(values, "run_id", out var runId) ||
-            !TryGetRequiredRelayValue(values, "step_id", out var stepId))
-        {
-            return false;
-        }
-
-        var metadata = new Dictionary<string, string>(StringComparer.Ordinal)
-        {
-            ["channel.platform"] = NormalizeOptional(message.Platform) ?? string.Empty,
-            ["channel.conversation_id"] = NormalizeOptional(message.Conversation?.PlatformId)
-                                          ?? NormalizeOptional(message.Conversation?.Id)
-                                          ?? string.Empty,
-        };
-        if (NormalizeOptional(message.MessageId) is { } messageId)
-            metadata["channel.message_id"] = messageId;
-
-        var approved = ResolveRelayApproved(values);
-        command = new WorkflowResumeCommand(
-            actorId,
-            runId,
-            stepId,
-            NormalizeOptional(message.MessageId),
-            approved,
-            ResolveRelayUserInput(values, approved),
-            metadata,
-            ResolveRelayEditedContent(values),
-            ResolveRelayFeedback(values, approved));
-        return true;
-    }
-
-    private static Dictionary<string, string> ExtractRelayCardActionValues(JsonElement root)
-    {
-        var values = new Dictionary<string, string>(StringComparer.Ordinal);
-        CopyRelayCardActionValues(root, values);
-
-        if (root.TryGetProperty("value", out var actionValue))
-            CopyRelayCardActionValues(actionValue, values);
-        if (root.TryGetProperty("form_value", out var formValue))
-            CopyRelayCardActionValues(formValue, values);
-
-        return values;
-    }
-
-    private static void CopyRelayCardActionValues(JsonElement element, IDictionary<string, string> values)
-    {
-        if (element.ValueKind != JsonValueKind.Object)
-            return;
-
-        foreach (var property in element.EnumerateObject())
-        {
-            switch (property.Value.ValueKind)
-            {
-                case JsonValueKind.String:
-                    values[property.Name] = property.Value.GetString() ?? string.Empty;
-                    break;
-                case JsonValueKind.Number:
-                case JsonValueKind.True:
-                case JsonValueKind.False:
-                    values[property.Name] = property.Value.ToString();
-                    break;
-            }
-        }
-    }
-
-    private static bool TryGetRequiredRelayValue(
-        IReadOnlyDictionary<string, string> values,
-        string key,
-        out string value)
-    {
-        value = string.Empty;
-        if (!values.TryGetValue(key, out var raw))
-            return false;
-
-        value = (raw ?? string.Empty).Trim();
-        return !string.IsNullOrWhiteSpace(value);
-    }
-
-    private static bool ResolveRelayApproved(IReadOnlyDictionary<string, string> values)
-    {
-        if (values.TryGetValue("approved", out var rawApproved) &&
-            bool.TryParse(rawApproved, out var approved))
-        {
-            return approved;
-        }
-
-        if (values.TryGetValue("action", out var rawAction))
-            return ResolveRelayDecisionLikeValue(rawAction);
-
-        if (values.TryGetValue("decision", out var rawDecision))
-            return ResolveRelayDecisionLikeValue(rawDecision);
-
-        return true;
-    }
-
-    private static bool ResolveRelayDecisionLikeValue(string? raw)
-    {
-        var normalized = (raw ?? string.Empty).Trim().ToLowerInvariant();
-        return normalized switch
-        {
-            "reject" or "rejected" or "deny" or "denied" or "cancel" => false,
-            _ => true,
-        };
-    }
-
-    private static string? ResolveRelayUserInput(
-        IReadOnlyDictionary<string, string> values,
-        bool approved)
-    {
-        var preferredKeys = approved
-            ? new[] { "edited_content", "user_input", "input", "comment" }
-            : new[] { "user_input", "comment", "input", "edited_content" };
-
-        foreach (var key in preferredKeys)
-        {
-            if (values.TryGetValue(key, out var raw))
-            {
-                var normalized = NormalizeOptional(raw);
-                if (normalized is not null)
-                    return normalized;
-            }
-        }
-
-        return null;
-    }
-
-    private static string? ResolveRelayEditedContent(IReadOnlyDictionary<string, string> values)
-    {
-        if (!values.TryGetValue("edited_content", out var raw))
-            return null;
-
-        return NormalizeOptional(raw);
-    }
-
-    private static string? ResolveRelayFeedback(
-        IReadOnlyDictionary<string, string> values,
-        bool approved)
-    {
-        if (approved)
-        {
-            if (values.TryGetValue("user_input", out var approvedRaw))
-                return NormalizeOptional(approvedRaw);
-
-            return null;
-        }
-
-        foreach (var key in new[] { "user_input", "comment", "input", "edited_content" })
-        {
-            if (values.TryGetValue(key, out var raw))
-            {
-                var normalized = NormalizeOptional(raw);
-                if (normalized is not null)
-                    return normalized;
-            }
-        }
-
-        return null;
-    }
-
-    private static string GetRelayContentType(RelayContent? content)
-    {
-        var normalizedContentType = NormalizeRelayContentType(content?.ContentType);
-        if (!string.IsNullOrWhiteSpace(normalizedContentType))
-            return normalizedContentType;
-
-        return NormalizeRelayContentType(content?.Type);
-    }
-
-    private static string NormalizeRelayContentType(string? contentType) =>
-        NormalizeOptional(contentType)?.ToLowerInvariant() ?? string.Empty;
-
-    private static string? NormalizeOptional(string? value)
-    {
-        var normalized = (value ?? string.Empty).Trim();
-        return normalized.Length == 0 ? null : normalized;
-    }
-
-    private static IResult FriendlyReply(string text) =>
-        Results.Json(new { reply = new { text } });
-
-    // ─── Relay payload models ───
-
-    private static readonly JsonSerializerOptions RelayJsonOptions = new()
-    {
-        PropertyNameCaseInsensitive = true,
-        PropertyNamingPolicy = JsonNamingPolicy.SnakeCaseLower,
-    };
-
-    private sealed class RelayMessage
-    {
-        public string? MessageId { get; set; }
-        public string? Platform { get; set; }
-        public RelayAgent? Agent { get; set; }
-        public RelayConversation? Conversation { get; set; }
-        public RelaySender? Sender { get; set; }
-        public RelayContent? Content { get; set; }
-        public string? Timestamp { get; set; }
-    }
-
-    private sealed class RelayAgent
-    {
-        public string? ApiKeyId { get; set; }
-        public string? Name { get; set; }
-    }
-
-    private sealed class RelayConversation
-    {
-        public string? Id { get; set; }
-        public string? PlatformId { get; set; }
-        public string? Type { get; set; }
-    }
-
-    private sealed class RelaySender
-    {
-        public string? PlatformId { get; set; }
-        public string? DisplayName { get; set; }
-    }
-
-    private sealed class RelayContent
-    {
-        public string? ContentType { get; set; }
-        public string? Type { get; set; }
-        public string? Text { get; set; }
-    }
-
-    private sealed class RelayReplyAccumulator
-    {
-        private readonly object _gate = new();
-        private readonly StringBuilder _responseBuilder = new();
-        private string? _errorMessage;
-        private bool _wasTruncated;
-
-        public RelayReplyAccumulator(int maxChars)
-        {
-            MaxChars = maxChars > 0 ? maxChars : 16 * 1024;
-        }
-
-        public int MaxChars { get; }
-
-        public bool WasTruncated
-        {
-            get
-            {
-                lock (_gate)
-                    return _wasTruncated;
-            }
-        }
-
-        public bool IsEmpty
-        {
-            get
-            {
-                lock (_gate)
-                    return _responseBuilder.Length == 0;
-            }
-        }
-
-        public void Append(string? text)
-        {
-            if (string.IsNullOrEmpty(text))
-                return;
-
-            lock (_gate)
-            {
-                if (_responseBuilder.Length >= MaxChars)
-                {
-                    _wasTruncated = true;
-                    return;
-                }
-
-                var remaining = MaxChars - _responseBuilder.Length;
-                if (text.Length <= remaining)
-                {
-                    _responseBuilder.Append(text);
-                    return;
-                }
-
-                _responseBuilder.Append(text, 0, remaining);
-                _wasTruncated = true;
-            }
-        }
-
-        public void SetError(string? errorMessage)
-        {
-            if (string.IsNullOrWhiteSpace(errorMessage))
-                return;
-
-            lock (_gate)
-                _errorMessage = errorMessage.Trim();
-        }
-
-        public string? GetErrorMessage()
-        {
-            lock (_gate)
-                return _errorMessage;
-        }
-
-        public string Snapshot()
-        {
-            lock (_gate)
-                return _responseBuilder.ToString();
-        }
-    }
-
 }
